moved swaggerDefinition from utils to new docs folder and changed the path in generateAPISchema file
